Refactor: modularize core application (app.c split into input, ui, scene, env)
